requirements.txt
Flask
requests
tensorflow
tensorflow-hub
numpy
pillow
